Separate code into several files
