+++ lib/datadog/statsd.rb
@@ -1,6 +1,11 @@
 # frozen_string_literal: true
 require 'socket'
 
+require_relative 'statsd/telemetry'
+require_relative 'statsd/udp_connection'
+require_relative 'statsd/uds_connection'
+require_relative 'statsd/batch'
+
 # = Datadog::Statsd: A DogStatsd client (https://www.datadoghq.com)
 #
 # @example Set up a global Statsd client for a server on localhost:8125
@@ -19,212 +24,6 @@ require 'socket'
 #   statsd = Datadog::Statsd.new 'localhost', 8125, tags: 'tag1:true'
 module Datadog
   class Statsd
-
-    class Telemetry
-      attr_accessor :metrics
-      attr_accessor :events
-      attr_accessor :service_checks
-      attr_accessor :bytes_sent
-      attr_accessor :bytes_dropped
-      attr_accessor :packets_sent
-      attr_accessor :packets_dropped
-      attr_reader   :estimate_max_size
-
-      def initialize(disabled, tags)
-        @disabled = disabled
-        @tags = tags
-        reset
-
-        # estimate_max_size is an estimation or the maximum size of the
-        # telemetry payload. Since we don't want our packet to go over
-        # 'max_buffer_bytes', we have to adjust with the size of the telemetry
-        # (and any tags used). The telemetry payload size will change depending
-        # on the actual value of metrics: metrics received, packet dropped,
-        # etc. This is why we add a 63bytes margin: 9 bytes for each of the 7
-        # telemetry metrics.
-        @estimate_max_size = @disabled ? 0 : flush().length + 9 * 7
-      end
-
-      def reset
-        @metrics = 0
-        @events = 0
-        @service_checks = 0
-        @bytes_sent = 0
-        @bytes_dropped = 0
-        @packets_sent = 0
-        @packets_dropped = 0
-      end
-
-      def flush
-        return '' if @disabled
-
-        # using shorthand syntax to reduce the garbage collection
-        return %Q(
-datadog.dogstatsd.client.metrics:#{@metrics}|#{COUNTER_TYPE}|##{@tags}
-datadog.dogstatsd.client.events:#{@events}|#{COUNTER_TYPE}|##{@tags}
-datadog.dogstatsd.client.service_checks:#{@service_checks}|#{COUNTER_TYPE}|##{@tags}
-datadog.dogstatsd.client.bytes_sent:#{@bytes_sent}|#{COUNTER_TYPE}|##{@tags}
-datadog.dogstatsd.client.bytes_dropped:#{@bytes_dropped}|#{COUNTER_TYPE}|##{@tags}
-datadog.dogstatsd.client.packets_sent:#{@packets_sent}|#{COUNTER_TYPE}|##{@tags}
-datadog.dogstatsd.client.packets_dropped:#{@packets_dropped}|#{COUNTER_TYPE}|##{@tags})
-      end
-    end
-
-    class Connection
-      DEFAULT_HOST = '127.0.0.1'
-      DEFAULT_PORT = 8125
-
-      # StatsD host. Defaults to 127.0.0.1.
-      attr_reader :host
-
-      # StatsD port. Defaults to 8125.
-      attr_reader :port
-
-      # DogStatsd unix socket path. Not used by default.
-      attr_reader :socket_path
-
-      def initialize(telemetry)
-        @telemetry = telemetry
-      end
-
-      # Close the underlying socket
-      def close
-        @socket && @socket.close
-      end
-
-      def write(message)
-        @logger.debug { "Statsd: #{message}" } if @logger
-        payload = message + @telemetry.flush()
-        send_message(payload)
-
-        @telemetry.reset
-        @telemetry.bytes_sent += payload.length
-        @telemetry.packets_sent += 1
-      rescue StandardError => boom
-        # Try once to reconnect if the socket has been closed
-        retries ||= 1
-        if retries <= 1 &&
-          (boom.is_a?(Errno::ENOTCONN) or
-           boom.is_a?(Errno::ECONNREFUSED) or
-           boom.is_a?(IOError) && boom.message =~ /closed stream/i)
-          retries += 1
-          begin
-            @socket = connect
-            retry
-          rescue StandardError => e
-            boom = e
-          end
-        end
-
-        @telemetry.bytes_dropped += payload.length
-        @telemetry.packets_dropped += 1
-        @logger.error { "Statsd: #{boom.class} #{boom}" } if @logger
-        nil
-      end
-
-      private
-
-      def socket
-        @socket ||= connect
-      end
-    end
-
-    class UDPConnection < Connection
-      def initialize(host, port, logger, telemetry)
-        super(telemetry)
-        @host = host || ENV.fetch('DD_AGENT_HOST', nil) || DEFAULT_HOST
-        @port = port || ENV.fetch('DD_DOGSTATSD_PORT', nil) || DEFAULT_PORT
-        @logger = logger
-      end
-
-      private
-
-      def connect
-        socket = UDPSocket.new
-        socket.connect(@host, @port)
-        socket
-      end
-
-      def send_message(message)
-        socket.send(message, 0)
-      end
-    end
-
-    class UDSConnection < Connection
-      class BadSocketError < StandardError; end
-
-      def initialize(socket_path, logger, telemetry)
-        super(telemetry)
-        @socket_path = socket_path
-        @logger = logger
-      end
-
-      private
-
-      def connect
-        socket = Socket.new(Socket::AF_UNIX, Socket::SOCK_DGRAM)
-        socket.connect(Socket.pack_sockaddr_un(@socket_path))
-        socket
-      end
-
-      def send_message(message)
-        socket.sendmsg_nonblock(message)
-      rescue Errno::ECONNREFUSED, Errno::ECONNRESET, Errno::ENOENT => e
-        @socket = nil
-        raise BadSocketError, "#{e.class}: #{e}"
-      end
-    end
-
-    class Batch
-      def initialize(connection, max_buffer_bytes)
-        @connection = connection
-        @max_buffer_bytes = max_buffer_bytes
-        @depth = 0
-        reset
-      end
-
-      def open
-        @depth += 1
-        yield
-      ensure
-        @depth -= 1
-        flush if !open?
-      end
-
-      def open?
-        @depth > 0
-      end
-
-      def add(message)
-        message_bytes = message.bytesize
-
-        unless @buffer_bytes == 0
-          if @buffer_bytes + 1 + message_bytes >= @max_buffer_bytes
-            flush
-          else
-            @buffer << NEW_LINE
-            @buffer_bytes += 1
-          end
-        end
-
-        @buffer << message
-        @buffer_bytes += message_bytes
-      end
-
-      def flush
-        return if @buffer_bytes == 0
-        @connection.write @buffer
-        reset
-      end
-
-      private
-
-      def reset
-        @buffer = String.new
-        @buffer_bytes = 0
-      end
-    end
-
     # Create a dictionary to assign a key to every parameter's name, except for tags (treated differently)
     # Goal: Simple and fast to add some other parameters
     OPTS_KEYS = {
